Move GetSecurityGroupsResponse to right utils package.
