+++ src/main/java/cloud/fogbow/ras/core/plugins/interoperability/openstack/publicip/v2/OpenStackPublicIpPlugin.java
@@ -6,7 +6,7 @@ import java.util.Properties;
 
 import cloud.fogbow.common.constants.OpenStackConstants;
 import cloud.fogbow.common.exceptions.InternalServerErrorException;
-import cloud.fogbow.ras.core.plugins.interoperability.openstack.util.serializables.v2.GetSecurityGroupsResponse;
+import cloud.fogbow.ras.core.plugins.interoperability.openstack.util.v2.serializables.GetSecurityGroupsResponse;
 import org.apache.log4j.Logger;
 
 import com.google.gson.JsonSyntaxException;
@@ -32,7 +32,7 @@ import cloud.fogbow.ras.core.plugins.interoperability.openstack.network.v2.Creat
 import cloud.fogbow.ras.core.plugins.interoperability.openstack.network.v2.RemoveSecurityGroupFromServerRequest;
 import cloud.fogbow.ras.core.plugins.interoperability.openstack.publicip.v2.GetFloatingIpResponse.FloatingIp;
 import cloud.fogbow.ras.core.plugins.interoperability.openstack.publicip.v2.GetNetworkPortsResponse.Port;
-import cloud.fogbow.ras.core.plugins.interoperability.openstack.util.serializables.v2.GetSecurityGroupsResponse.SecurityGroup;
+import cloud.fogbow.ras.core.plugins.interoperability.openstack.util.v2.serializables.GetSecurityGroupsResponse.SecurityGroup;
 
 public class OpenStackPublicIpPlugin implements PublicIpPlugin<OpenStackV3User> {
