+++ src/main/java/cloud/fogbow/ras/core/plugins/interoperability/openstack/util/v2/OpenStackCloudUtils.java
@@ -2,7 +2,8 @@ package cloud.fogbow.ras.core.plugins.interoperability.openstack.util.v2;
 
 import cloud.fogbow.common.exceptions.InternalServerErrorException;
 import cloud.fogbow.ras.constants.Messages;
-import cloud.fogbow.ras.core.plugins.interoperability.openstack.util.serializables.v2.GetSecurityGroupsResponse;
+import cloud.fogbow.ras.core.plugins.interoperability.openstack.util.v2.serializables.GetSecurityGroupsResponse;
+import cloud.fogbow.ras.core.plugins.interoperability.openstack.util.v2.serializables.GetSecurityGroupsResponse.SecurityGroup;
 import org.apache.log4j.Logger;
 import org.json.JSONException;
 
@@ -12,7 +13,7 @@ public class OpenStackCloudUtils {
 
     public static String getSecurityGroupIdFromGetResponse(String json) throws InternalServerErrorException {
         try {
-            GetSecurityGroupsResponse.SecurityGroup securityGroup = GetSecurityGroupsResponse.fromJson(json).getSecurityGroups().iterator().next();
+            SecurityGroup securityGroup = GetSecurityGroupsResponse.fromJson(json).getSecurityGroups().iterator().next();
             return securityGroup.getId();
         } catch (JSONException e) {
             LOGGER.error(String.format(Messages.Log.UNABLE_TO_RETRIEVE_NETWORK_ID_S, json), e);
